refactor: :art: Move User from  users.schemas to users.models and delete users.schemas
